requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.71.1
google-cloud-documentai==2.25.0
google-cloud-storage==2.15.0
vertexai==1.71.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0
PyPDF2==3.0.1
python-magic==0.4.27
pdf2image==1.17.0
openpyxl==3.1.2
reportlab==4.0.7